Restructure
This commit does a number of things:
1. Gash commands no longer need sourcing, they are files in their
   own right, small-scripts run by kaem.
2. Hence kaem has been significantly improved.
3. tinycc has been seperated into two stages. The resulting tcc
   is now a "pure" upstream tcc 0.9.27, rather than the mes
   0.9.26+patches.
4. Absolute paths based on ${REPO} are now used.
5. As per gash commands a gash generation stsage has been added.
6. bashbug has been fixed, instead of patching it out.
7. A dummy mes-loader (which dosen't really exist) has been added
   to tinycc options to remove a bug (same as what guix does).

This differs to GNU Guix in a number of important ways.

1. Guix does not have relative gash-core-utils. It is locked into
   /. This causes non-portability and does not work for
   mescc-tools-seed.
2. Guix doesn't build the final tcc until much later, just before
   the first GNU triplet. Why? I'm not sure. The 0.9.26+patches
   tinycc is rather buggy and was occassionally producing a segfaulting
   make. This tinycc produces much better binaries, and is much
   more stable and modern.

The one thing that does relate to GNU Guix is the mes-loader business.
I believe this is partly responsible to the fix of make.

Going forward, I will look at doing coreutils next, as I dislike this
slow gash-core-utils setup quite a lot - it can be rather buggy. Once
again this is different to Guix, they don't do coreutils until just
before the GNU toolchain.