Move to tarball-based setup for everything post-gash
This creates the setup to be much more simple. It dosen't require
me to create a thousand git repositories, it is the standard, and
when we get to network access - this is by far the simplest way
to do it, instead of porting git.

Hence, this also adds patch, to allow patches to be added.

This also adds prompts for the user to populate the REPO variable.